Relocate coords. Run ruff.
